refactor(tests): move ScopedTempDir to shared header with pid+counter unique names
- Add tests/Platform/ScopedTempDir.h (Platform::Test::ScopedTempDir) with a
  private uniqueName() helper that appends both pid and an atomic counter,
  guaranteeing uniqueness within a single process even during coverage runs
  where all tests execute in one process
- Remove the pre-clean remove_all() from the constructor; the unique name makes
  it unnecessary
- Replace the duplicate struct definitions in test_LinuxSystemProbe.cpp and
  test_LinuxProcessProbe.cpp with an include + using declaration
- Remove <format>, <system_error>, <unistd.h> from test_LinuxSystemProbe.cpp
  since they were only needed by the now-moved ScopedTempDir
