+++ tests/Platform/test_LinuxSystemProbe.cpp
@@ -9,19 +9,16 @@
 #if defined(__linux__) && __has_include(<unistd.h>)
 
 #include "Platform/Linux/LinuxSystemProbe.h"
+#include "Platform/ScopedTempDir.h"
 #include "Platform/SystemTypes.h"
 
 #include <atomic>
 #include <chrono>
 #include <filesystem>
-#include <format>
 #include <fstream>
-#include <system_error>
 #include <thread>
 #include <unordered_map>
 #include <vector>
-
-#include <unistd.h>
 
 namespace Platform
 {
@@ -568,31 +565,7 @@ TEST(LinuxSystemProbeTest, CacheConcurrentAccessIsSafe)
 
 // ========== Error Path / Injection Tests ==========
 
-// RAII helper: creates a uniquely-named temp directory (with PID suffix to
-// avoid parallel-test collisions) and removes it on destruction so cleanup
-// happens even if an assertion fires.
-struct ScopedTempDir
-{
-    std::filesystem::path path;
-
-    explicit ScopedTempDir(std::string_view name) : path(std::filesystem::temp_directory_path() / std::format("{}_{}", name, getpid()))
-    {
-        std::error_code ec;
-        std::filesystem::remove_all(path, ec); // best-effort pre-clean; ignore error
-        std::filesystem::create_directories(path);
-    }
-
-    ~ScopedTempDir() noexcept
-    {
-        std::error_code ec;
-        std::filesystem::remove_all(path, ec); // must not throw from destructor
-    }
-
-    ScopedTempDir(const ScopedTempDir&) = delete;
-    ScopedTempDir& operator=(const ScopedTempDir&) = delete;
-    ScopedTempDir(ScopedTempDir&&) = delete;
-    ScopedTempDir& operator=(ScopedTempDir&&) = delete;
-};
+using Platform::Test::ScopedTempDir;
 
 TEST(LinuxSystemProbeTest, MissingStatFileSilentlyReturnsZeroCpu)
 {
